addressed the comments, move the contents from benchmarks to parent directory
