--- integration_tests/benchmarks/.gitignore
@@ -1,7 +1,0 @@
-tfjs-converter/
-tfjs-core/
-tfjs-data/
-tfjs-layers/
-tfjs-node/
-data/
-coverage/
