+++ integration_tests/types.ts
@@ -320,6 +320,7 @@ export interface TensorDef {
   shape: number[];
   dtype: 'string'|'float32'|'int32'|'bool'|'complex64';
 }
+
 export interface ValidationRun {
   taskType: 'model';
   modelFormat: ModelFormat;
